# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Hub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResult {
  success: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  uid: string;
  name: string;
  email: string;
  grade: number;
  role: UserRole;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  user: "user",
  admin: "admin",
} as const;

export interface CreateUserBody {
  uid: string;
  name: string;
  email: string;
  grade: number;
  role?: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  user: "user",
  admin: "admin",
} as const;

export interface UpdateUserBody {
  name?: string;
  grade?: number;
  role?: UpdateUserBodyRole;
}

export type UserStatsByGrade = { [key: string]: number };

export interface UserStats {
  total: number;
  byGrade: UserStatsByGrade;
  admins: number;
  newToday: number;
}

export type NoteContentType =
  (typeof NoteContentType)[keyof typeof NoteContentType];

export const NoteContentType = {
  text: "text",
  pdf: "pdf",
  image: "image",
} as const;

export interface Note {
  id: number;
  grade: number;
  subject: string;
  chapter: string;
  title: string;
  contentType: NoteContentType;
  content: string;
  createdAt: string;
}

export type CreateNoteBodyContentType =
  (typeof CreateNoteBodyContentType)[keyof typeof CreateNoteBodyContentType];

export const CreateNoteBodyContentType = {
  text: "text",
  pdf: "pdf",
  image: "image",
} as const;

export interface CreateNoteBody {
  grade: number;
  subject: string;
  chapter: string;
  title: string;
  contentType: CreateNoteBodyContentType;
  content: string;
}

export type UpdateNoteBodyContentType =
  (typeof UpdateNoteBodyContentType)[keyof typeof UpdateNoteBodyContentType];

export const UpdateNoteBodyContentType = {
  text: "text",
  pdf: "pdf",
  image: "image",
} as const;

export interface UpdateNoteBody {
  grade?: number;
  subject?: string;
  chapter?: string;
  title?: string;
  contentType?: UpdateNoteBodyContentType;
  content?: string;
}

export type McqCorrectAnswer =
  (typeof McqCorrectAnswer)[keyof typeof McqCorrectAnswer];

export const McqCorrectAnswer = {
  A: "A",
  B: "B",
  C: "C",
  D: "D",
} as const;

export type McqDifficulty = (typeof McqDifficulty)[keyof typeof McqDifficulty];

export const McqDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Mcq {
  id: number;
  grade: number;
  subject: string;
  chapter: string;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: McqCorrectAnswer;
  difficulty: McqDifficulty;
  createdAt: string;
}

export type CreateMcqBodyCorrectAnswer =
  (typeof CreateMcqBodyCorrectAnswer)[keyof typeof CreateMcqBodyCorrectAnswer];

export const CreateMcqBodyCorrectAnswer = {
  A: "A",
  B: "B",
  C: "C",
  D: "D",
} as const;

export type CreateMcqBodyDifficulty =
  (typeof CreateMcqBodyDifficulty)[keyof typeof CreateMcqBodyDifficulty];

export const CreateMcqBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface CreateMcqBody {
  grade: number;
  subject: string;
  chapter: string;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: CreateMcqBodyCorrectAnswer;
  difficulty: CreateMcqBodyDifficulty;
}

export type UpdateMcqBodyCorrectAnswer =
  (typeof UpdateMcqBodyCorrectAnswer)[keyof typeof UpdateMcqBodyCorrectAnswer];

export const UpdateMcqBodyCorrectAnswer = {
  A: "A",
  B: "B",
  C: "C",
  D: "D",
} as const;

export type UpdateMcqBodyDifficulty =
  (typeof UpdateMcqBodyDifficulty)[keyof typeof UpdateMcqBodyDifficulty];

export const UpdateMcqBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface UpdateMcqBody {
  grade?: number;
  subject?: string;
  chapter?: string;
  question?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer?: UpdateMcqBodyCorrectAnswer;
  difficulty?: UpdateMcqBodyDifficulty;
}

export interface Pyq {
  id: number;
  grade: number;
  subject: string;
  title: string;
  year: number;
  pdfUrl: string;
  createdAt: string;
}

export interface CreatePyqBody {
  grade: number;
  subject: string;
  title: string;
  year: number;
  pdfUrl: string;
}

export interface Task {
  id: number;
  uid: string;
  text: string;
  completed: boolean;
  createdAt: string;
}

export interface CreateTaskBody {
  uid: string;
  text: string;
}

export interface UpdateTaskBody {
  text?: string;
  completed?: boolean;
}

export interface ScoreEntry {
  id: number;
  uid: string;
  userName: string;
  grade: number;
  score: number;
  totalQuestions: number;
  subject: string;
  createdAt: string;
}

export interface SubmitScoreBody {
  uid: string;
  userName: string;
  grade: number;
  score: number;
  totalQuestions: number;
  subject: string;
}

export type AiChatBodyHistoryItemRole =
  (typeof AiChatBodyHistoryItemRole)[keyof typeof AiChatBodyHistoryItemRole];

export const AiChatBodyHistoryItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export type AiChatBodyHistoryItem = {
  role: AiChatBodyHistoryItemRole;
  content: string;
};

export interface AiChatBody {
  message: string;
  history?: AiChatBodyHistoryItem[];
}

export interface AiChatResponse {
  reply: string;
}

export type ListNotesParams = {
  grade?: number;
  subject?: string;
  chapter?: string;
};

export type ListNoteSubjectsParams = {
  grade?: number;
};

export type ListMcqsParams = {
  grade?: number;
  subject?: string;
  chapter?: string;
  difficulty?: string;
  limit?: number;
};

export type ListPyqsParams = {
  grade?: number;
  subject?: string;
};

export type ListTasksParams = {
  uid: string;
};

export type ListScoresParams = {
  period?: ListScoresPeriod;
};

export type ListScoresPeriod =
  (typeof ListScoresPeriod)[keyof typeof ListScoresPeriod];

export const ListScoresPeriod = {
  daily: "daily",
  alltime: "alltime",
} as const;