# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Student Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes with optional filters
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
        - name: subject
          in: query
          schema:
            type: string
        - name: chapter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /notes/subjects:
    get:
      operationId: listNoteSubjects
      tags: [notes]
      summary: List distinct subjects for a grade
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Subjects list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /mcqs:
    get:
      operationId: listMcqs
      tags: [mcqs]
      summary: List MCQs with filters
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
        - name: subject
          in: query
          schema:
            type: string
        - name: chapter
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of MCQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mcq"
    post:
      operationId: createMcq
      tags: [mcqs]
      summary: Create MCQ (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMcqBody"
      responses:
        "201":
          description: MCQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mcq"
  /mcqs/{id}:
    patch:
      operationId: updateMcq
      tags: [mcqs]
      summary: Update MCQ (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMcqBody"
      responses:
        "200":
          description: MCQ updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mcq"
    delete:
      operationId: deleteMcq
      tags: [mcqs]
      summary: Delete MCQ (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: MCQ deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /pyqs:
    get:
      operationId: listPyqs
      tags: [pyqs]
      summary: List previous year questions
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
        - name: subject
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of PYQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pyq"
    post:
      operationId: createPyq
      tags: [pyqs]
      summary: Create PYQ (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePyqBody"
      responses:
        "201":
          description: PYQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pyq"
  /pyqs/{id}:
    delete:
      operationId: deletePyq
      tags: [pyqs]
      summary: Delete PYQ (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PYQ deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /scores:
    get:
      operationId: listScores
      tags: [scores]
      summary: Leaderboard scores
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, alltime]
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreEntry"
    post:
      operationId: submitScore
      tags: [scores]
      summary: Submit a quiz score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoreBody"
      responses:
        "201":
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreEntry"
...
  /ai/chat:
    post:
      operationId: nepAiChat
      tags: [ai]
      summary: Send a message to Nep AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    Note:
      type: object
      properties:
        id:
          type: integer
        grade:
          type: integer
        subject:
          type: string
        chapter:
          type: string
        title:
          type: string
        contentType:
          type: string
          enum: [text, pdf, image]
        content:
          type: string
        createdAt:
          type: string
      required: [id, grade, subject, chapter, title, contentType, content, createdAt]
    CreateNoteBody:
      type: object
      properties:
        grade:
          type: integer
        subject:
          type: string
        chapter:
          type: string
        title:
          type: string
        contentType:
          type: string
          enum: [text, pdf, image]
        content:
          type: string
      required: [grade, subject, chapter, title, contentType, content]
    UpdateNoteBody:
      type: object
      properties:
        grade:
          type: integer
        subject:
          type: string
        chapter:
          type: string
        title:
          type: string
        contentType:
          type: string
          enum: [text, pdf, image]
        content:
          type: string
    Mcq:
      type: object
      properties:
        id:
          type: integer
        grade:
          type: integer
        subject:
          type: string
        chapter:
          type: string
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        createdAt:
          type: string
      required: [id, grade, subject, chapter, question, optionA, optionB, optionC, optionD, correctAnswer, difficulty, createdAt]
    CreateMcqBody:
      type: object
      properties:
        grade:
          type: integer
        subject:
          type: string
        chapter:
          type: string
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D]
        difficulty:
          type: string
          enum: [easy, medium, hard]
      required: [grade, subject, chapter, question, optionA, optionB, optionC, optionD, correctAnswer, difficulty]
    UpdateMcqBody:
      type: object
      properties:
        grade:
          type: integer
        subject:
          type: string
        chapter:
          type: string
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D]
        difficulty:
          type: string
          enum: [easy, medium, hard]
    Pyq:
      type: object
      properties:
        id:
          type: integer
        grade:
          type: integer
        subject:
          type: string
        title:
          type: string
        year:
          type: integer
        pdfUrl:
          type: string
        createdAt:
          type: string
      required: [id, grade, subject, title, year, pdfUrl, createdAt]
    CreatePyqBody:
      type: object
      properties:
        grade:
          type: integer
        subject:
          type: string
        title:
          type: string
        year:
          type: integer
        pdfUrl:
          type: string
      required: [grade, subject, title, year, pdfUrl]
...
    ScoreEntry:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        userName:
          type: string
        grade:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        subject:
          type: string
        createdAt:
          type: string
      required: [id, uid, userName, grade, score, totalQuestions, subject, createdAt]
    SubmitScoreBody:
      type: object
      properties:
        uid:
          type: string
        userName:
          type: string
        grade:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        subject:
          type: string
      required: [uid, userName, grade, score, totalQuestions, subject]
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
            required: [role, content]
      required: [message]
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
      required: [reply]